+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Smaller initial blobs.
 - Player blob slightly bigger than most drone initial size.
 - Larger span of initial drone sizes.
+- Generate random vectors with perlin noise instead of UP/DOWN/LEFT/RIGHT.
 
 ## [0.7.0] - 2016-12-19
 ### Added